python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ofc"
version = "0.1.0"
description = "Open Floor Control - Compose and run multi-agent teams"
readme = "../README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OFC Contributors"}
]
keywords = ["ai", "agents", "multi-agent", "llm", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.scripts]
ofc = "ofc.cli:main"

[project.urls]
Homepage = "https://ofc.dev"
Repository = "https://github.com/ofc/ofc"

[tool.setuptools.packages.find]
include = ["ofc*", "chatroom*"]
